requirements.txt
openai-whisper==20250625
python-docx==1.1.2

